LayoutTests/js/dom/dfg-custom-getter-expected.txt

Tests that DFG custom getter caching does not break the world.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo() is " DIV #text DIV #text SCRIPT" on all iterations including after DFG tier-up.
PASS successfullyParsed is true

TEST COMPLETE
Some stuff
More stuff
